// src/db/generated/models/EducationEntry.ts
export type EducationEntryModel = runtime.Types.Result.DefaultSelection<Prisma.$EducationEntryPayload>
// ...
export type $EducationEntryPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "EducationEntry"
  objects: {
    caregiver: Prisma.$CaregiverPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    agencyId: string
    caregiverId: string
    schoolName: string
    city: string | null
    state: string | null
    programOrDegree: string | null
    completedOn: Date | null
    graduated: boolean | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["educationEntry"]>
  composites: {}
}