// src/db/generated/models/EmergencyContact.ts
export type EmergencyContactModel = runtime.Types.Result.DefaultSelection<Prisma.$EmergencyContactPayload>
// ...
export type $EmergencyContactPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "EmergencyContact"
  objects: {
    caregiver: Prisma.$CaregiverPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    agencyId: string
    caregiverId: string
    fullName: string
    relationship: string | null
    phone: string | null
    alternatePhone: string | null
    email: string | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["emergencyContact"]>
  composites: {}
}